+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ## [Unreleased]
 
+### Added
+* RasterIter struct
+* Raster::region, region_iter, set_region methods
+* Gray, Mask, Rgb now implement Iterator (for use with set_region)
 ### Changed
-* Fixed bugs in channel conversions.
+* Fixed bugs in channel conversions
+* Region::intersection, right, bottom reworked
+### Removed
+* Raster::set_rect (use set_region instead)
 
 ## [0.4.0] - 2019-05-11
 ### Added